🧪 improved the AI modal for Chores
 ✨Now at 50% accuracy

+++ Backend/src/main/java/UniNest/Backend/controller/ChoreController.java
@@ -79,7 +79,7 @@ public class ChoreController {
         predictionRequest.setRoom(chore.getRoom());
         predictionRequest.setDifficultyScore(chore.getDifficultyScore());
         predictionRequest.setEstDurationMin(chore.getEstDurationMin());
-        predictionRequest.setFrequencyPerWeek(chore.getFrequencyPerWeek());
+        predictionRequest.setFrequencyPerWeek((double) chore.getFrequencyPerWeek());
 
         //  predict assignment
         ChorePredictionResponse assignment =
